refactor: complete migration of aarch64 suite
